requirements-dev.txt
flake8==3.5.0
nose-exclude
coverage
coveralls
json2html